auth_schemas.py
from pydantic import BaseModel

class LoginRequest(BaseModel):
    username: str
    password: str

class SignupRequest(BaseModel):
    username: str
    email: str
    password: str

class LoginResponse(BaseModel):
    access_token: str
    user_id: int
    message: str

class SignupResponse(BaseModel):
    message: str