requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
requests
stripe
websockets
alpaca-py
passlib[bcrypt]
python-jose[cryptography]
bcrypt>=4.0.0